# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "My API",
        "version": "1.0.0"
    },
    "tags": [
        {
            "name": "Article"
        }
    ],
    "paths": {
        "/article": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Retourne tous les articles",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Article"
                ],
                "summary": "Créer un article",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article créé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/article/{articleId}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Retourne un article selon son identifiant d'article",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "informatique"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "Modifie un article selon son identifiant d'article",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "informatique"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Article modifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/article/id/{id}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Retourne un article selon son identifiant mongo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "63370cb0ea103eb2c5d51aee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Article"
                ],
                "summary": "Supprime un article selon son identifiant mongo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "63370cb0ea103eb2c5d51aee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "deleted": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/Article"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Article"
                ],
                "summary": "Modifie un article selon son identifiant mongo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "63370cb0ea103eb2c5d51aee"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Article"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Article"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/article/tag/{tag}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Retourne tous les articles selon un tag",
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "informatique"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Article"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/article/markdown/{language}/{articleId}": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Retourne un article au format markdown selon son identifiant d'article",
                "parameters": [
                    {
                        "name": "language",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "fr"
                        }
                    },
                    {
                        "name": "articleId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "informatique"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "markdown": {
                                            "type": "string",
                                            "example": "### Bonjour monde! Ceci est un test."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Article": {
                "type": "object",
                "properties": {
                    "articleIdentifier": {
                        "type": "string",
                        "example": "informatique"
                    },
                    "author": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "views": {
                        "type": "integer",
                        "example": 0
                    },
                    "isVerified": {
                        "type": "boolean",
                        "example": false
                    },
                    "filename": {
                        "type": "string",
                        "example": "article.md"
                    },
                    "defaultLanguage": {
                        "type": "string",
                        "example": "fr"
                    },
                    "created": {
                        "type": "string",
                        "example": "2020-01-01T00:00:00.000Z"
                    },
                    "updated": {
                        "type": "string",
                        "example": "2020-01-01T00:00:00.000Z"
                    },
                    "name": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "example": {
                            "fr": "Informatique",
                            "en": "Computer Science"
                        }
                    },
                    "description": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "string"
                        },
                        "example": {
                            "fr": "Informatique description",
                            "en": "Computer Science description"
                        }
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "informatique",
                            "computer science"
                        ]
                    },
                    "sources": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Wikipedia"
                                },
                                "url": {
                                    "type": "string",
                                    "example": "https://en.wikipedia.org/wiki/Computer_science"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}